Add the functions Get_vrt() and Create_chips()
Get_vrt(), Create_chips() and their descriptions make the code
more readable to the user. The functions are called within a for loop
that loops over multiple dated directories each containing a .vrt file.
The .vrt file for a specific date is obtained and Create_chips() uses
this file and a provided text file containing corner coordinates
for glaciers needing to be chipped. The chips are saved within the
dated directory.

The chips provide a time series of the changes of glacier position.
However, glacier chips for several positions are saved within a
directory for a single date in time. Instead, we want a directory for a
certain glacier that holds multiple glacier chips for the same position,
but varying over time.